==> 1_2_wrangling.Rmd <==
---
title: "project"
author: "David Hyon"
date: "2025-04-26"
output: pdf_document
---

# Load packages

```{r}
library(tigris)
library(tidyverse)
library(sf)
library(tidycensus)
library(lubridate)
library(rlang)

## Define functions
pad_zero <- function(x) {
  x %>%
    as.character() %>%
    if_else(nchar(.) == 1, paste0("0", .), .)
}

summarise_changes <- function(data, column, tag) {
  column_sym <- ensym(column)         # capture column name as symbol
  column_name <- tag # get the string name
  
  data %>%
    summarise(
      periods = n(),
      !!paste0("final_", column_name) := last(!!column_sym, na_rm = TRUE),
      !!paste0("average_", column_name) := mean(!!column_sym, na.rm = TRUE),
      !!paste0("sd_", column_name) := sd(!!column_sym, na.rm = TRUE),
      !!paste0("positive_", column_name) := sum(!!column_sym > 0, na.rm = TRUE),
      !!paste0("negative_", column_name) := sum(!!column_sym < 0, na.rm = TRUE),
      !!paste0("max_increase_", column_name) := max(!!column_sym, na.rm = TRUE),
      !!paste0("max_decrease_", column_name) := min(!!column_sym, na.rm = TRUE),
      .groups = "drop"
    )
}


add_state_fips <- function(df, state_col) {
  
  # Lookup table for states and their FIPS codes
  state_fips_lookup <- data.frame(
    state = c("Alabama", "Alaska", "Arizona", "Arkansas", "California",
              "Colorado", "Connecticut", "Delaware", "Florida", "Georgia",
              "Hawaii", "Idaho", "Illinois", "Indiana", "Iowa", "Kansas",
              "Kentucky", "Louisiana", "Maine", "Maryland", "Massachusetts",
              "Michigan", "Minnesota", "Mississippi", "Missouri", "Montana",
              "Nebraska", "Nevada", "New Hampshire", "New Jersey", "New Mexico",
              "New York", "North Carolina", "North Dakota", "Ohio", "Oklahoma",
              "Oregon", "Pennsylvania", "Rhode Island", "South Carolina",
              "South Dakota", "Tennessee", "Texas", "Utah", "Vermont", "Virginia",
              "Washington", "West Virginia", "Wisconsin", "Wyoming"),
    fips = c("01", "02", "04", "05", "06",
             "08", "09", "10", "12", "13",
             "15", "16", "17", "18", "19", "20",
             "21", "22", "23", "24", "25",
             "26", "27", "28", "29", "30",
             "31", "32", "33", "34", "35",
             "36", "37", "38", "39", "40",
             "41", "42", "44", "45",
             "46", "47", "48", "49", "50", "51",
             "53", "54", "55", "56")
  )
  
  # Merge to add FIPS code
  df <- df %>%
    left_join(state_fips_lookup, by = setNames("state", state_col)) %>% 
    mutate(
      statefips = as.numeric(fips),
      .keep = "unused"
    )
  
  return(df)
}





```




# Gameplan

# Election

## Read in 2024

```{r}
electionRaw <- read_csv("./2024_US_County_Level_Presidential_Results.csv")


election24 <- electionRaw %>% 
  select(county_fips, 1:9) %>% 
  mutate(
    trump = case_when(per_gop > per_dem ~ 1,
                      per_dem > per_gop ~ 0,
                      per_gop == per_dem ~ -999,
                      .default = NA),
    bell = if_else(trump == 1, 1, 0)
  )

table(election24$trump, useNA = c("always"))
rm(electionRaw)
```

## Read in 2020

```{r}
electionRaw <- read_csv("./2020_US_County_Level_Presidential_Results.csv")


election20 <- electionRaw %>% 
   select(county_fips, 1:9) %>% # Keeps state and county name
  #select(county_fips, 4:9) %>%  # Only voting data
  mutate(
    trump = case_when(per_gop > per_dem ~ 1,
                      per_dem > per_gop ~ 0,
                      per_gop == per_dem ~ -999,
                      .default = NA),
    bell = if_else(trump == 0, 1, 0)
  )

table(election20$trump, useNA = c("always"))
rm(electionRaw)
str(election20)
```
## Read in 2016

```{r}
electionRaw <- read_csv("./2016_US_County_Level_Presidential_Results.csv")


election16 <- electionRaw %>% 
  mutate(
    county_fips = combined_fips
  ) %>% 
  select(county_fips, 2:7) %>% 
  mutate(
    trump = case_when(per_gop > per_dem ~ 1,
                      per_dem > per_gop ~ 0,
                      per_gop == per_dem ~ -999,
                      .default = NA),
    bell = if_else(trump == 1, 1, 0)
  )

table(election16$trump, useNA = c("always"))

rm(electionRaw)
str(election16)
```

## Clean and Combine

```{r}

# Deal with special case for Connecticut which changed from counties to Planning regions in 2024




election <- election24 %>% 
  left_join(election20, by = join_by(county_fips), suffix = c("_24", "")) %>%
  mutate(
    GEOID = county_fips,
    county_fips = as.numeric(county_fips)
  ) %>% 
  left_join(election16, by = join_by(county_fips), suffix = c("_20", "_16")) %>% 
  mutate(
    bellweather = ifelse(bell_16 == 1 & bell_20 == 1 & bell_24 == 1, 1, 0),
    pop_vote_bell = ifelse(trump_16 == 0 & trump_20 == 0 & trump_24 == 1, 0, 1),
    bell_20_24 = ifelse(bell_20 == 1 & bell_24 == 1, 1, 0),
    change = ifelse(trump_16 == trump_20 & trump_20 == trump_24, 0, 1),
    change_20_24 = ifelse(trump_20 == trump_24, 0, 1)
  ) %>% 
  filter(!(state_name %in% c("Alaska","Connecticut", "District of Columbia"))) %>% 
  add_state_fips("state_name")

table(election$change)
table(election$bellweather)
table(election$pop_vote_bell)

#rm(election16, election20, election24)

```
```{r}
write_csv(election, "./cleanData/election.csv")
```





## Covid

There are 12 Datasets available with the COVID data

1. Consumer Spending

### County

```{r}
spendingRaw <- read_csv("./EconomicTracker-main/data/Affinity - County - Daily.csv")

#str(spendingCounty)

#spendingNational <- read_csv("./EconomicTracker-main/data/Affinity - National - Daily.csv")

#str(spendingNational)

spendingRaw %>% 
  #filter(year == 2022) %>% 
  mutate(
    spend_all = ifelse(spend_all == ".", NA, spend_all),
    missing = ifelse(is.na(spend_all), 1, 0)
  ) %>% 
  group_by(countyfips) %>% 
  filter(countyfips==1001)

```
```{r}



spend <- spendingRaw %>% 
  mutate(
    spend_all = ifelse(spend_all == ".", NA, spend_all),
    missing = ifelse(is.na(spend_all), 1, 0)
  ) %>% 
  mutate(
    month = pad_zero(month),
    day = pad_zero(day)
  ) %>% 
  unite("date", year, month, day, sep = "", remove = FALSE) %>% 
  mutate(
    date = ymd(date),
    woy = week(date),
    dow = wday((date))
  ) %>% 
  filter(dow==1)%>% 
  mutate(
    spend_all = as.numeric(spend_all)
  ) 
# Summarize by countyfips
spendingFinal <- spend %>%
  filter(!is.na(spend_all)) %>% 
  group_by(countyfips) %>%
  mutate(
    check = n()
  ) %>% 
  filter(check > 200) %>% 
  summarise_changes(spend_all, "consumer_spending")


write_csv(spendingFinal,"./cleanData/consumerSpending.csv")  

  
```

### State

```{r}
spendingRaw <- read_csv("./EconomicTracker-main/data/Affinity - State - Daily.csv")

spendState <- spendingRaw %>% 
  mutate(
    spend_all = ifelse(spend_all == ".", NA, spend_all),
    missing = ifelse(is.na(spend_all), 1, 0)
  ) %>% 
  mutate(
    month = pad_zero(month),
    day = pad_zero(day)
  ) %>% 
  unite("date", year, month, day, sep = "", remove = FALSE) %>% 
  mutate(
    date = ymd(date),
    woy = week(date),
    dow = wday((date))
  ) %>% 
  filter(dow==1)%>% 
  mutate(
    spend_all = as.numeric(spend_all)
  ) 
# Summarize by statefips
spendingStateFinal <- spendState %>%
  filter(!is.na(spend_all)) %>% 
  group_by(statefips) %>%
  # mutate(
  #   check = n()
  # ) %>% 
  # filter(check > 200) %>% 
  summarise_changes(spend_all, "consumer_spending")


```


2. Small Business

### County

```{r}
businessRaw <- read_csv("./EconomicTracker-main/data/Womply - County - Weekly.csv")
```

```{r}
sbOpen <- businessRaw %>% 
  group_by(countyfips) %>% 
  summarise_changes(merchants_all, "sb_open")

sbRevenue <- businessRaw %>% 
  group_by(countyfips) %>% 
  summarise_changes(revenue_all, "sb_rev")



filter(businessRaw, countyfips == 20057)

filter(sbOpen, countyfips ==20057)


```


### State

```{r}
businessRaw <- read_csv("./EconomicTracker-main/data/Womply - State - Weekly.csv")

sbOpenState <- businessRaw %>% 
  group_by(statefips) %>% 
  summarise_changes(merchants_all, "sb_open")

sbRevenueState <- businessRaw %>% 
  group_by(statefips) %>% 
  summarise_changes(revenue_all, "sb_rev")

```



## Employment

### County

```{r}

employmentRaw <- read_csv("./EconomicTracker-main/data/Employment - County - Weekly.csv")
```
```{r}

  
employmentRaw[employmentRaw == "."] <- NA


employment <- employmentRaw %>% 
  filter(!is.na(emp)) %>% 
  filter(year < 2025) %>% 
  filter(!(year == 2024 & month >= 11)) %>% 
  mutate(
    emp = as.numeric(emp)
  )


employmentClean <- employment %>% 
  group_by(countyfips) %>% 
  summarise_changes(emp, "employment")

```
### State

```{r}
employmentRaw <- read_csv("./EconomicTracker-main/data/Employment - State - Weekly.csv")

employmentRaw[employmentRaw == "."] <- NA


employmentState <- employmentRaw %>% 
  filter(!is.na(emp)) %>% 
  filter(year < 2025) %>% 
  filter(!(year == 2024 & month >= 11)) %>% 
  mutate(
    emp = as.numeric(emp)
  )


employmentStateClean <- employmentState %>% 
  group_by(statefips) %>% 
  summarise_changes(emp, "employment")

```


## Unemployment

### County

```{r}

uiCountyRaw <- read_csv("./EconomicTracker-main/data/UI Claims - County - Weekly.csv")

uiCountyRaw[uiCountyRaw == "."] <- NA 

uiCounty <- uiCountyRaw %>% 
  filter(!is.na(initclaims_rate_regular))%>% 
  filter(year < 2025) %>% 
  filter(!(year == 2024 & month >= 11)) %>% 
  mutate(
    initclaims_rate_regular = as.numeric(initclaims_rate_regular)
  )
  
uiCountyClean <- uiCounty %>% 
  group_by(countyfips) %>% 
  summarise_changes(initclaims_rate_regular, "ui_claims_rate") %>% 
  select(-6:-7)




```
### State

```{r}

uiStateRaw <- read_csv("./EconomicTracker-main/data/UI Claims - State - Weekly.csv")

uiStateRaw[uiStateRaw == "."] <- NA 

uiState <- uiStateRaw %>% 
  filter(!is.na(initclaims_rate_regular))%>% 
  filter(year < 2025) %>% 
  filter(!(year == 2024 & month >= 11)) %>% 
  mutate(
    initclaims_rate_regular = as.numeric(initclaims_rate_regular)
  )
  
uiStateClean <- uiState %>% 
  group_by(statefips) %>% 
  summarise_changes(initclaims_rate_regular, "ui_claims_rate") %>% 
  select(-6:-7)



```





```{r}
covidRaw1 <- read_csv("./EconomicTracker-main/data/COVID - County - Daily 2020.csv")

covidRaw2 <- read_csv("./EconomicTracker-main/data/COVID - County - Daily 2021.csv")

covidRaw3 <- read_csv("./EconomicTracker-main/data/COVID - County - Daily 2022.csv")

covidRaw <- covidRaw1 %>% 
  bind_rows(covidRaw2, covidRaw3)


covidRaw[covidRaw == "."] <- NA 


covid <- covidRaw %>% 
  mutate(
    month = pad_zero(month),
    day = pad_zero(day)
  ) %>% 
  unite("date", year, month, day, sep = "", remove = FALSE) %>% 
  mutate(
    date = ymd(date),
    woy = week(date),
    dow = wday((date))
  ) %>% 
  filter(dow==1) %>% 
  mutate(across(5:29, as.numeric))


covid  %>% 
  filter(countyfips==6057)
  #filter(!is.na(case_count))


covid %>%
  
  filter(countyfips==6057) %>% str()
  #filter(case_count>0)


covidClean <- covid %>% 
  group_by(countyfips) %>% 
  summarize(
    final_death_rate = max(death_rate, na.rm=TRUE),
    final_case_rate = max(case_rate, na.rm=TRUE),
    max_new_death_rate = max(new_death_rate, na.rm = TRUE),
    max_new_case_rate = max(new_case_rate, na.rm = TRUE),
    weeks = sum(!is.na(case_count)),
    avg_new_case_rate = sum(new_case_rate, na.rm = TRUE) / weeks,
    avg_new_death_rate = sum(new_death_rate, na.rm = TRUE) / weeks,
  ) %>% 
  mutate(across(2:8, ~ifelse(weeks==0, 0, .x)))

```
## Google Mobility

### County

```{r}
googleRaw <- read_csv("./EconomicTracker-main/data/Google Mobility - County - Daily.csv")


google <- googleRaw %>% 
  mutate(
    month = pad_zero(month),
    day = pad_zero(day)
  ) %>% 
  unite("date", year, month, day, sep = "", remove = FALSE) %>% 
  mutate(
    date = ymd(date),
    woy = week(date),
    dow = wday((date))
  ) %>% 
  filter(date >= ymd("2020-03-01") & date <= ymd("2022-12-31"))


google %>% 
  filter(countyfips==1023)
#google[google == "."] <- NA 

googleClean <- google %>% 
  group_by(countyfips) %>% 
  mutate(
    google = as.numeric(gps_away_from_home)
  ) %>% 
  filter(!is.na(google)) %>% 
  summarise_changes(google, "away_from_home") %>% 
  mutate(
    percent_positive = positive_away_from_home/ (positive_away_from_home + negative_away_from_home),
    .keep = "unused"
  )

```
### State

```{r}
googleRaw <- read_csv("./EconomicTracker-main/data/Google Mobility - State - Daily.csv")


googleState <- googleRaw %>% 
  mutate(
    month = pad_zero(month),
    day = pad_zero(day)
  ) %>% 
  unite("date", year, month, day, sep = "", remove = FALSE) %>% 
  mutate(
    date = ymd(date),
    woy = week(date),
    dow = wday((date))
  ) %>% 
  filter(date >= ymd("2020-03-01") & date <= ymd("2022-12-31"))


googleStateClean <- googleState %>% 
  group_by(statefips) %>% 
  mutate(
    google = as.numeric(gps_away_from_home)
  ) %>% 
  filter(!is.na(google)) %>% 
  summarise_changes(google, "away_from_home") %>% 
  mutate(
    percent_positive = positive_away_from_home/ (positive_away_from_home + negative_away_from_home),
    .keep = "unused"
  )
```


# Join all datasets

```{r}

countyData <- election %>% 
  left_join(covidClean, by =join_by(county_fips==countyfips)) %>% 
  left_join(employmentClean, by =join_by(county_fips==countyfips)) %>%
  left_join(googleClean, by =join_by(county_fips==countyfips)) %>% 
  left_join(sbOpen, by =join_by(county_fips==countyfips)) %>% 
  left_join(sbRevenue, by =join_by(county_fips==countyfips)) %>% 
  left_join(spendingFinal, by =join_by(county_fips==countyfips)) %>% 
  left_join(uiCountyClean, by =join_by(county_fips==countyfips)) %>% 
  select(-starts_with("periods"))



# ## Imputing Missing Data
# 
# just use state, and fill in for missing counties, use the same variables as available for county data. This 
# makes it an apples to apples comparison. For swing states / bellwether counties, explore this in EDA, just
# use classification models for predicting 2024 across country. 

joinCountyState <- function(electDF, dfCounty, dfState) {
  
  elect_county <- semi_join(electDF, dfCounty, by =join_by(county_fips==countyfips)) 
  elect_county2 <- left_join(elect_county, dfCounty, by =join_by(county_fips==countyfips)) 
  
  elect_state1 <- anti_join(electDF, dfCounty, by =join_by(county_fips==countyfips))
  elect_state2 <- left_join(elect_state1, dfState, by = join_by(statefips))
  
  final_df <- bind_rows(elect_county2, elect_state2)
  
  #return(elect_county2)
  return(final_df)
  
  
}





countyDataImputed <- election %>% 
  left_join(covidClean, by =join_by(county_fips==countyfips)) %>%
  joinCountyState(employmentClean, employmentStateClean) %>% 
  joinCountyState(googleClean, googleStateClean) %>%
  joinCountyState(sbOpen, sbOpenState) %>%
  joinCountyState(sbRevenue, sbRevenueState) %>%
  joinCountyState(spendingFinal, spendingStateFinal) %>%
  joinCountyState(uiCountyClean, uiStateClean) 
  


```











## Demographic

Get 2010 and 2020, as well as ACS
- Age and Sex
```{r}
ageSexRaw <- read_csv("./census/ageSex/ACSDT5Y2020.B01001-Data.csv", skip = 1)

ageSex <- ageSexRaw %>%  
  mutate(
    county_fips = as.numeric(str_replace(Geography, ".*US", ""))
  ) %>% 
  select(-starts_with("Margin")) %>% janitor::clean_names() %>%
  mutate(
    across(
      starts_with("estimate_total_") & !matches("estimate_total$"),
      ~ (.x / estimate_total) * 100
    )
  ) %>%
  # Step 2: Remove total columns
  select(
    -estimate_total,
    -estimate_total_male,
    -estimate_total_female
  )%>%
  select(-geography, -geographic_area_name, -x101)


```



- Race 

```{r}
raceRaw <- read_csv("./census/race/DECENNIALPL2020.P1-Data.csv", skip = 1) %>% 
  janitor::clean_names() %>% 
  select(1:11)

race <- raceRaw %>%  
  # Step 1: Remove total_population_of_one_race
  select(-total_population_of_one_race) %>%
  
  # Step 2: Convert remaining columns (except geography, name, total) to percentages
  mutate(
    across(
      .cols = -c(geography, geographic_area_name, total),
      .fns = ~ (.x / total) * 100
      #.names = "{.col}_percent"
    )
  ) %>% 
  # Step 3: Clean column names: remove "total_population_of_" from the new names
  rename_with(
    ~ str_remove(., "^total_population_of_"),
    .cols = -c(geography, geographic_area_name)
  ) %>% 
  mutate(
    county_fips = as.numeric(str_replace(geography, ".*US", ""))
  ) %>%
  select(-geography, -geographic_area_name, -total)

```



- Ethnicity

```{r}
ethnicityRaw <- read_csv("./census/ethnicity/DECENNIALPL2020.P2-Data.csv", skip = 1) %>% 
  janitor::clean_names() %>% 
  select(1:4)


ethnicity <- ethnicityRaw %>% 
  mutate(
    across(
      .cols = -c(geography, geographic_area_name, total),
      .fns = ~ (.x / total) * 100
      #.names = "{.col}_percent"
    )
  )  %>% 
  mutate(
    county_fips = as.numeric(str_replace(geography, ".*US", ""))
  ) %>%
  select(-geography, -geographic_area_name, -total)

```


- Education 

```{r}
educationRaw <- read_csv("./census/education/ACSDT5Y2020.B15003-Data.csv", skip =1) %>% 
  select(-starts_with("Margin")) %>% janitor::clean_names() %>% 
  select(1:27)

education <- educationRaw %>% mutate(
    across(
      starts_with("estimate_total_") & !matches("estimate_total$"),
      ~ (.x / estimate_total) * 100
    )
  )  %>% 
  mutate(
    county_fips = as.numeric(str_replace(geography, ".*US", ""))
  ) %>%
  select(-geography, -geographic_area_name, -estimate_total)
  
```



- Income 

```{r}
incomeRaw <- read_csv("./census/income/ACSST5Y2020.S1901-Data.csv", skip =1) %>% 
  select(-starts_with("Margin")) %>% janitor::clean_names() %>% 
  select(1:15)


income <- incomeRaw %>% 
  mutate(
    county_fips = as.numeric(str_replace(geography, ".*US", ""))
  ) %>%
  select(-geography, -geographic_area_name, -estimate_households_total)
```



- Population 

```{r}
popRaw <- read_csv("./census/population/DECENNIALDHC2020.H8-Data.csv", skip = 1)


pop <- popRaw %>% 
  mutate(
    county_fips = as.numeric(str_replace(Geography, ".*US", "")),
    population = `!!Total`
  ) %>% 
  select(county_fips, population)

countyDataImputed %>% 
  left_join(pop, by = join_by(county_fips)) %>% 
  filter(is.na(population))


```

## Join to COVID data

```{r}

master <- countyDataImputed %>% 
  left_join(ageSex, by = join_by(county_fips)) %>% 
  left_join(race, by = join_by(county_fips)) %>% 
  left_join(ethnicity, by = join_by(county_fips)) %>% 
  left_join(education, by = join_by(county_fips)) %>% 
  left_join(income, by = join_by(county_fips)) %>% 
  left_join(pop, by = join_by(county_fips)) %>% 
  select(-starts_with("periods"))
  
colnames(master) 
  
write_csv(master, "./cleanData/master.csv")

```
